+++ src/main/java/fr/neocraft/main/util/Vector3f.java
@@ -7,8 +7,8 @@ import net.minecraft.entity.Entity;
 public class Vector3f implements Serializable{
 
 	private static final long serialVersionUID = 546456451117L;
-	public int x,y,z;
-	public Vector3f(int X,int Y,int Z) {
+	public float x,y,z;
+	public Vector3f(float X,float Y,float Z) {
 		x=X;
 		y=Y;
 		z=Z;
@@ -23,7 +23,7 @@ public class Vector3f implements Serializable{
 		return v2.x == x & v2.y == y & v2.z == z;
 	}
 	
-	public Vector6f createSquare(int u, int v, int w)
+	public Vector6f createSquare(float u, float v, float w)
 	{
 		return new Vector6f(x-u/2, y - v/2, z - w/2, u/2, v/2, w/2);
 	}
@@ -31,6 +31,6 @@ public class Vector3f implements Serializable{
 	
 	public static Vector3f getVectorFromEntity(Entity p)
 	{
-		return new Vector3f((int)p.posX,(int) p.posY, (int)p.posZ);
+		return new Vector3f((float)p.posX,(float) p.posY, (float)p.posZ);
 	}
 }
